# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lecture Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication and user profile
  - name: videos
    description: Lecture video management
  - name: storage
    description: Object storage upload and serving endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      description: Returns the current authenticated user's profile including role
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List videos
      description: Returns published videos for viewers, all videos for admins
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Filter by title
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter by category
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Create a new video (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Video created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /videos/stats:
    get:
      operationId: getVideoStats
      tags: [videos]
      summary: Get video statistics (admin only)
      responses:
        "200":
          description: Video statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      operationId: updateVideo
      tags: [videos]
      summary: Update video metadata (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "200":
          description: Video updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Video deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /videos/{id}/status:
    patch:
      operationId: updateVideoStatus
      tags: [videos]
      summary: Publish or unpublish a video (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /videos/{id}/stream-url:
    get:
      operationId: getVideoStreamUrl
      tags: [videos]
      summary: Get a signed URL to stream a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Signed stream URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamUrlResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve a stored object (video file)
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, viewer]
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - role
        - createdAt

    Video:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        lecturer:
          type: string
        category:
          type: string
        videoUrl:
          type: string
        status:
          type: string
          enum: [draft, published]
        createdAt:
          type: string
        updatedAt:
          type: string
        uploadedBy:
          type: string
      required:
        - id
        - title
        - description
        - lecturer
        - category
        - videoUrl
        - status
        - createdAt
        - updatedAt

    VideoInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        lecturer:
          type: string
        category:
          type: string
        videoUrl:
          type: string
        status:
          type: string
          enum: [draft, published]
      required:
        - title
        - description
        - lecturer
        - category
        - status

    VideoStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [draft, published]
      required:
        - status

    VideoStats:
      type: object
      properties:
        total:
          type: integer
        published:
          type: integer
        draft:
          type: integer
        categories:
          type: array
          items:
            type: string
      required:
        - total
        - published
        - draft
        - categories

    StreamUrlResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url

    UploadUrlRequest:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType

    UploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath

    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required:
        - error